services/device/requirements.txt
flask==3.1.3
pyserial==3.5
